requirements-python.txt
ansible>=2.9.16,<2.10

# needed for inventory and aws modules
boto3

# required by ansible ec2 module
boto

# needed for python `ipaddr` filter
ipaddress
netaddr

# needed for managing databases (e.g. creating project db in infra step)
# mysql
PyMySQL

# needed for running docker (e.g. building aws lambda artifacts)
docker-py

# some tasks call aws command on localhost
awscli

# required for awscli
PyYAML<5.3

# required for `cs.ansible-plugins` role
jq

# required if host running ansible have selinux installed
selinux

# required for error pages
libsass